Change link pointing to unsupported Kubernetes docs
The README contained a link pointing to the docs of Kubernetes v1.18.
The docs for that version are no longer online, and when visiting the
link your browser shows a message about an invalid certificate.

In this commit, I changed the file to point to a newer version of the docs.