fix: use image of container in MigratorSpec
Previous to this commit, the operator would inject an waiter container
that would poll for the image of the first container in the target Pod
rather than the image of the specified container. This is surprising
behaviour and also potentially problematic when combined with Istio's
`holdApplicationUntilProxyStart` functionality, which injects the Istio
sidecar at the beginning of the list of containers.

This commit ensures the operator uses the container specified in the
MigratorSpec, if supplied and present. If the target container name is
specified and not present, it falls back to using the image from the
first container in the Pod (the previous behaviour). If the target
container name is not present, it also uses the previous behaviour.