fix: make findOwners handle custom types better
This change stops the controller from bubbling up an error when it
encounters a Kind that is has not registered in its scheme or an object
that it does not have access to. In these situations, the controller
will now stop climbing the ownership tree instead.

The new behaviour is desireable when an unsupported type (such as one
from a company's in-house operator) is marked as the controller of the
resource that controls the pods themselves.